changed to uint8_t

+++ includes/block.hpp
@@ -10,5 +10,5 @@ public:
 	int getType(void);
 	void setType(int t);
 private:
-	int type = AIR_BLOCK; //need to change to smaller data size ie. char/short// 1==grass, 0==air, 2==sand, 3==snow 
+	uint8_t type = AIR_BLOCK; //need to change to smaller data size ie. char/short// 1==grass, 0==air, 2==sand, 3==snow 
 };
